Venv and PIP/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.2.2
urllib3==2.0.7
